Tests: Cleanup spacing, added 1 test, formatting
The amount of spacing used here is beyond me :) Adapted a new .editorconfig
rule to apply for spacing, formatting etc.

Added: a simple property test for name of chat.

Removed: (Commented Out) the imports for CallRoute.js and Department.js since
these do not exist as yet in the list of classes.

Author: Michael Dabydeen <mdabydeen@gmail.com>
Signed-Off-By: Michael Dabydeen <michael@skrumble.com>